pyproject.toml
[project]
name = "bid-tool-v2"
version = "2.0.0"
description = "入札情報収集ツール V2 — 官公需API + e-Tokyo → Google Spreadsheet"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "python-dotenv>=1.0.0",
    "defusedxml>=0.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "bandit>=1.7.0",
]

[project.scripts]
bidtool = "src.main:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
